+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Add experimental support for Kotlin Value types as columns in the KotlinMapper.
+
 # 3.49.3
 
 - Only use JFR if the flight recorder was registered (#2736)